Move 'vectorize_function' to the 'vectorization' module, update tests.
